[New] Added test cases for pages dashboard
