+++ app/javascript/components/pages/dashboard/utils/to-reporting-location-table.js
@@ -9,7 +9,7 @@ import { INDICATOR_NAMES } from "../constants";
 const reportingLocationLabel = (reportingLocationConfig, i18n) => {
   const locationTypes = [];
 
-  reportingLocationConfig.label_keys.forEach(key => {
+  reportingLocationConfig?.label_keys?.forEach(key => {
     locationTypes.push(`${i18n.t(`location.base_types.${key}`)}`);
   });
